have separate module for confirm project deletion. have subfolder under modals/ for deletion modals
