+++ src/controllers/Screen/displayPage/displayTasks.js
@@ -2,7 +2,7 @@ import { format, isToday, isPast } from "date-fns";
 import displayTaskDetailsInModal from "../modals/displayTaskDetailsInModal";
 import { createAppController } from "../../App/createAppController";
 import displayPage from ".";
-import handleDelete from "../modals/handleDelete";
+import handleDelete from "../modals/deletion/handleDelete";
 
 const SVGS = {
   dueDate: `<svg xmlns="http://www.w3.org/2000/svg" width="17" height="17" viewBox="0 0 24 24" fill="none"
