+++ src/controllers/Screen/modals/deletion/confirmProjectDeletion.js
@@ -1,0 +1,3 @@
+export default function confirmProjectDeletion(projectObject) {
+  alert("deleting");
+}
